// src/types/api.ts
export interface Project {
  id?: string;
  uuid?: string;
  project_id?: string;
  project_uuid?: string;
  created_at?: string;
  updated_at?: string;
  name?: string;
  description?: string;
  custom_instruction?: string;
  instruction?: {
    content?: string;
  };
  owner?: number;
  team?: number;
  tags?: string[];
  starred_by?: number[];
  icon?: LucideIcon
  color?: string
  starred?: boolean
  lastUpdated?: string
  teamSize?: number
  chatCount?: number
  chatIcon?: LucideIcon
}